// devtools-ui/src/api/model/availableModuleFeature.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Awilixify Devtools API
 * API for inspecting module graphs, traces, and provider impact
 * OpenAPI spec version: 1.0.0
 */
import type { LifetimeType } from "./lifetimeType";
import type { ModuleFeatureOrigin } from "./moduleFeatureOrigin";

export interface AvailableModuleFeature {
	className: string;
	decoratorNames: string[];
	exported: boolean;
	key: string;
	moduleName: string;
	origin: ModuleFeatureOrigin;
	lifetime?: LifetimeType;
}